Move getInitialQueryContents to separate file
